+++ src/rwltypedefs.h
@@ -13,6 +13,7 @@
  *
  * History
  *
+ * bengsig 29-jun-2022 generate project
  * bengsig 04-mar-2022 printf project
  * bengsig 21-jan-2021 add rwl_pooltyp
  * bengsig 06-oct-2020 Remove sharding
@@ -43,6 +44,7 @@ typedef struct rwl_localvar rwl_localvar;
 typedef struct rwl_oerstat rwl_oerstat;
 typedef struct rwl_idlist rwl_idlist;
 typedef struct rwl_arglist rwl_arglist;
+typedef struct rwl_helplist rwl_helplist;
 typedef struct rwl_pathlist rwl_pathlist;
 typedef struct rwl_conlist rwl_conlist;
 typedef enum rwl_vsalloc rwl_vsalloc;